delete vscode and comment evm provider
