--- apps/mobile/.vscode/launch.json
@@ -1,13 +1,0 @@
-{
-  "configurations": [
-    {
-      "name": "Debug Expo Web - Experimental",
-      "request": "launch",
-      "type": "reactnativedirect",
-      "cwd": "${workspaceFolder}",
-      "platform": "expoweb",
-      "browserTarget": "chrome",
-      "url": "http://localhost:8081"
-    }
-  ]
-}
