+++ apps/mobile/app.json
@@ -66,6 +66,5 @@
       ]
     ],
     "sdkVersion": "51.0.0"
-  },
-  "name": "mobile"
+  }
 }
