+++ apps/mobile/src/app/Wrapper.tsx
@@ -56,11 +56,11 @@ export const Wrapper: React.FC = () => {
                 <PortalizeProvider>
                   <DialogProvider>
                     <StarknetProvider>
-                      <EVMProvider>
+                      {/* <EVMProvider> */}
                         <ModalProviders>
                           <App />
                         </ModalProviders>
-                      </EVMProvider>
+                      {/* </EVMProvider> */}
                     </StarknetProvider>
                   </DialogProvider>
                 </PortalizeProvider>
